+++ projects/wvr-elements/src/lib/core/manifest/manifest-entry-request.ts
@@ -8,6 +8,8 @@ export interface ManifestEntryRequest {
   method?: RequestMethod;
   body?: any;
   options?: RequestOptions;
+  decode?: string[];
+  bodyHttpParams?: 'fromObject' | 'fromString';
   onSuccess?: Array<Action>;
   onFailure?: Array<Action>;
 }