+++ projects/wvr-elements/src/lib/core/modal/modal.ts
@@ -1,5 +1,3 @@
-import { ThemeVariant } from '../../shared/theme';
-
 export interface Modal {
   name: string;
   open: boolean;