+++ projects/wvr-elements/src/lib/core/rest/request.ts
@@ -9,4 +9,6 @@ export interface Request {
   body?: any;
   options?: RequestOptions;
   map?: map;
+  decode?: string[];
+  bodyHttpParams?: 'fromObject' | 'fromString';
 }